+++ nvim/lua/custom/plugins/markdown-ui.lua
@@ -77,4 +77,15 @@ return {
       })
     end,
   },
+
+  -- Browser preview with live sync scrolling, KaTeX, Mermaid, PlantUML
+  {
+    'iamcco/markdown-preview.nvim',
+    cmd = { 'MarkdownPreviewToggle', 'MarkdownPreview', 'MarkdownPreviewStop' },
+    ft = { 'markdown' },
+    build = 'cd app && npx --yes yarn install',
+    keys = {
+      { '<leader>mp', '<cmd>MarkdownPreviewToggle<CR>', desc = 'Preview in browser' },
+    },
+  },
 }